+++ server/src/entities/errors/authorization_error.ts
