+++ server/src/services/role_validator.ts
@@ -1,6 +1,6 @@
 import { Role } from "../entities/role";
 import { User } from "../entities/user";
 
-export function auth_RoleValidator(user: User, roles: Role[]): boolean {
+export function roleValidator(user: User, roles: Role[]): boolean {
     return roles.includes(user.role);
 }
