+++ server/src/services/user_service.ts
@@ -2,9 +2,9 @@ import "reflect-metadata";
 import { inject, injectable, named } from "inversify";
 import { UserRepository } from "../repositories/user_repository";
 import { User } from "../entities/user";
-import { auth_RoleValidator } from "./auth_RoleValidator";
+import { roleValidator } from "./role_validator";
 import { ROLES } from "../entities/role";
-import { AuthorizationError } from "../entities/Error/AuthorizationError";
+import { AuthorizationError } from "../entities/errors/authorization_error";
 
 @injectable()
 export class UserService {
@@ -16,7 +16,7 @@ export class UserService {
 
     async findUserByUserId(userId: number): Promise<User> {
         const user = await this.userRepository.findUserByUserId(userId);
-        if (!auth_RoleValidator(user, ROLES)) {
+        if (!roleValidator(user, ROLES)) {
             throw new AuthorizationError();
         }
